requirements.txt
transformers
pytorch-metric-learning
pyyaml
munch